refactor: divide `DeployList`
